# sql/distsqlrun: plumb context into RowSource.Next