--- a/pkg/ccl/sqlccl/csv.go
+++ b/pkg/ccl/sqlccl/csv.go
@@ -1424,7 +1424,7 @@ func (sp *sstWriter) Run(ctx context.Context, wg *sync.WaitGroup) {
 		batch := store.NewBatchWriter()
 		var key, val []byte
 		for {
-			row, err := input.NextRow()
+			row, err := input.NextRow(ctx)
 			if err != nil {
 				return err
 			}