--- a/pkg/sql/distsqlrun/distinct.go
+++ b/pkg/sql/distsqlrun/distinct.go
@@ -94,7 +94,7 @@ func (d *distinct) Run(ctx context.Context, wg *sync.WaitGroup) {
 func (d *distinct) mainLoop(ctx context.Context) (earlyExit bool, _ error) {
 	var scratch []byte
 	for {
-		row, meta := d.input.Next()
+		row, meta := d.input.Next(ctx)
 		if !meta.Empty() {
 			if meta.Err != nil {
 				return false, meta.Err