This is a prerequisite for making processors implement RowSource because
we will want to pass a context from Processor.Run to RowSource.Next in
order to maintain tracing data.

Release note: None