@RaduBerinde 2017-12-14
Any perf difference? (e.g. in `BenchmarkInfrastructure`)