@petermattis 2017-12-14
The diff seems to be noise (there is no way this PR should have made anything faster):

```
name                        old time/op  new time/op  delta
Infrastructure/n1/r1-8       174µs ± 4%   174µs ± 2%    ~     (p=0.481 n=10+10)
Infrastructure/n1/r100-8     298µs ± 1%   295µs ± 1%  -0.93%  (p=0.002 n=9+9)
Infrastructure/n1/r10000-8  14.4ms ± 2%  14.8ms ± 4%    ~     (p=0.315 n=9+10)
Infrastructure/n3/r1-8       565µs ± 4%   562µs ± 4%    ~     (p=0.481 n=10+10)
Infrastructure/n3/r100-8    1.38ms ± 2%  1.35ms ± 2%  -1.62%  (p=0.017 n=9+10)
Infrastructure/n3/r10000-8  54.5ms ± 2%  53.6ms ± 2%  -1.68%  (p=0.015 n=10+10)
```