@andreimatei 2017-12-14
Does this change actually use any of the plumbed contexts anywhere?
How do you see the future - if we make processors call each other, would we configure each one to either work as they do now (with a `Run()` method and an embedded context) or through calls to `Next()` that takes an explicit context? Like, processors would support both?
I'm wondering if we could/should make the `Next()` implementations also use embedded contexts - after all, the call graph is statically determined for flows, so we know in advance what context each processor should use.

---

Review status: 0 of 25 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20711#-:-L0KnrGqDBZUVekHjSo7:b-dchyz0)*
<!-- Sent from Reviewable.io -->