@petermattis 2017-12-14
> Does this change actually use any of the plumbed contexts anywhere?

No.

> How do you see the future - if we make processors call each other, would we configure each one to either work as they do now (with a Run() method and an embedded context) or through calls to Next() that takes an explicit context? Like, processors would support both?

As an interim step, yes. And then I imagine a future step where we replace `Processor.Run` with `runProcessor` that is common across all processors.

> I'm wondering if we could/should make the Next() implementations also use embedded contexts - after all, the call graph is statically determined for flows, so we know in advance what context each processor should use.

Won't that muck with the tracing data? I'd expect all of the "fused" processors to exist in a single span.

---

Review status: 0 of 25 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20711#-:-L0KphEuw87I6d_rSfhO:b-93wa16)*
<!-- Sent from Reviewable.io -->