@petermattis 2017-12-14
> If that's what we want, we can do that by embedding the same context in all the fused processors.
But I'm not sure that it is what we want. I think it'd be more useful for traces to continue looking like they look today - with processors having different spans. I think that whether or not communication between processors is synchronous should not affect how the trace looks.

Ok. Would this involve plumbing a context to processor creation? Let's discuss in person.

---

Review status: 0 of 25 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20711#-:-L0L6_EjOW7Mc-Japj7o:b-el93pj)*
<!-- Sent from Reviewable.io -->