@petermattis 2017-12-28
https://github.com/cockroachdb/cockroach/pull/20584 takes the approach of adding a `context.Context` to `FlowCtx` which is already plumbed to all processors. This should fulfill @andreimatei's desire that tracing is not affected by how processors are fused or not fused.