@arjunravinarayan 2018-01-02
Reviewed 25 of 25 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20711)*
<!-- Sent from Reviewable.io -->